django_/requirements.txt
asgiref==3.8.1
Django==5.1.2
PyMySQL==1.1.1
sqlparse==0.5.1
tzdata==2024.2
djangorestframework==3.15.2
